PB- newest update on search. Almost done
