+++ client/src/screens/Home/Home.jsx
@@ -3,9 +3,9 @@ import Layout from "../../components/shared/Layout/Layout";
 import Carousel from '../../components/Carousel/Carousel'
 import WelcomeContainer from '../../components/WelcomeContainer/WelcomeContainer'
 
-const Home = () => {
+const Home = (props) => {
   return (
-    <Layout>
+    <Layout handleChange={props.handleChange}>
       <h1>HomePage</h1>
       <Carousel />
       <WelcomeContainer/>
